+++ CircuitBreaker/CircuitBreakerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace RedirectionIO\Client\ProxySymfony\CircuitBreaker;
+
+use Symfony\Component\HttpFoundation\Request;
+
+interface CircuitBreakerInterface
+{
+    public function shouldNotProcessRequest(Request $request): bool;
+}