+++ DependencyInjection/Configuration.php
@@ -5,7 +5,10 @@ namespace RedirectionIO\Client\ProxySymfony\DependencyInjection;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
-class Configuration implements ConfigurationInterface
+/**
+ * @internal
+ */
+final class Configuration implements ConfigurationInterface
 {
     /**
      * {@inheritdoc}