+++ RedirectionIOBundle.php
@@ -2,8 +2,14 @@
 
 namespace RedirectionIO\Client\ProxySymfony;
 
+use RedirectionIO\Client\ProxySymfony\DependencyInjection\CompilerPass\CircuitBreakerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class RedirectionIOBundle extends Bundle
+final class RedirectionIOBundle extends Bundle
 {
+    public function build(ContainerBuilder $container)
+    {
+        $container->addCompilerPass(new CircuitBreakerPass());
+    }
 }